Share correlation estimation across replicates
Allows the option to consider the case where prevalence is simultaneously being estimated in n replicate surveys in different populations. Each replicate is identical (assumed survey design and design prevalence). Estimates of prevalence are separate, but ICC assumed to be the same for clusters within each survey. This allows sharing of the burden of estimating ICC across multiple surveys. For replicates = Inf this corresponds to perfect knowledge of ICC which is also useful as a limiting case.

To implement I've created a new sample variance function (`sv`) which takes the FI matrix and the number of replicates and returns the sample variance of one example prevalence estimate from the replicates and ICC.